requirements.txt
Flask
requests
google-cloud-datastore
fbchat